+++ sdk/core/azure-core/src/http/http_sanitizer.cpp
@@ -6,7 +6,6 @@
 #include "azure/core/url.hpp"
 
 #include <iterator>
-#include <regex>
 #include <sstream>
 
 namespace {