+++ sdk/eventhubs/azure-messaging-eventhubs/src/consumer_client.cpp
@@ -97,7 +97,9 @@ namespace Azure { namespace Messaging { namespace EventHubs {
     // Set the user agent related properties in the connectOptions based on the package
     // information and application ID.
     _detail::EventHubsUtilities::SetUserAgent(
-        connectOptions, m_consumerClientOptions.ApplicationID);
+        connectOptions,
+        m_consumerClientOptions.ApplicationID,
+        m_consumerClientOptions.CppStandardVersion);
 
     return Azure::Core::Amqp::_internal::Connection{
         m_fullyQualifiedNamespace, m_credential, connectOptions};