+++ sdk/identity/azure-identity/src/azure_pipelines_credential.cpp
@@ -3,10 +3,10 @@
 
 #include "azure/identity/azure_pipelines_credential.hpp"
 
+#include "private/client_assertion_credential_impl.hpp"
 #include "private/identity_log.hpp"
 #include "private/package_version.hpp"
 #include "private/tenant_id_resolver.hpp"
-#include "private/token_credential_impl.hpp"
 
 #include <azure/core/internal/json/json.hpp>
 
@@ -28,30 +28,6 @@ using Azure::Core::Json::_internal::json;
 using Azure::Identity::_detail::IdentityLog;
 using Azure::Identity::_detail::PackageVersion;
 using Azure::Identity::_detail::TenantIdResolver;
-using Azure::Identity::_detail::TokenCredentialImpl;
-
-namespace {
-bool IsValidTenantId(std::string const& tenantId)
-{
-  const std::string allowedChars = ".-";
-  if (tenantId.empty())
-  {
-    return false;
-  }
-  for (auto const c : tenantId)
-  {
-    if (allowedChars.find(c) != std::string::npos)
-    {
-      continue;
-    }
-    if (!StringExtensions::IsAlphaNumeric(c))
-    {
-      return false;
-    }
-  }
-  return true;
-}
-} // namespace
 
 AzurePipelinesCredential::AzurePipelinesCredential(
     std::string tenantId,
@@ -60,27 +36,22 @@ AzurePipelinesCredential::AzurePipelinesCredential(
     std::string systemAccessToken,
     AzurePipelinesCredentialOptions const& options)
     : TokenCredential("AzurePipelinesCredential"), m_serviceConnectionId(serviceConnectionId),
-      m_systemAccessToken(systemAccessToken),
-      m_clientCredentialCore(tenantId, options.AuthorityHost, options.AdditionallyAllowedTenants),
-      m_httpPipeline(HttpPipeline(options, "identity", PackageVersion::ToString(), {}, {}))
+      m_systemAccessToken(systemAccessToken)
 {
+  // Allow these headers to be logged since they are used for troubleshooting.
+  AzurePipelinesCredentialOptions optionsWithLoggableHeaders = options;
+  optionsWithLoggableHeaders.Log.AllowedHttpHeaders.insert("x-vss-e2eid");
+  optionsWithLoggableHeaders.Log.AllowedHttpHeaders.insert("x-msedge-ref");
+
+  m_httpPipeline = std::make_unique<HttpPipeline>(
+      optionsWithLoggableHeaders,
+      "identity",
+      PackageVersion::ToString(),
+      std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>>{},
+      std::vector<std::unique_ptr<Azure::Core::Http::Policies::HttpPolicy>>{});
+
   m_oidcRequestUrl = _detail::DefaultOptionValues::GetOidcRequestUrl();
 
-  bool isTenantIdValid = IsValidTenantId(tenantId);
-  if (!isTenantIdValid)
-  {
-    IdentityLog::Write(
-        IdentityLog::Level::Warning,
-        "Invalid tenant ID provided  for " + GetCredentialName()
-            + ". The tenant ID must be a non-empty string containing only alphanumeric characters, "
-              "periods, or hyphens. You can locate your tenant ID by following the instructions "
-              "listed here: https://learn.microsoft.com/partner-center/find-ids-and-domain-names");
-  }
-  if (clientId.empty())
-  {
-    IdentityLog::Write(
-        IdentityLog::Level::Warning, "No client ID specified for " + GetCredentialName() + ".");
-  }
   if (serviceConnectionId.empty())
   {
     IdentityLog::Write(
@@ -101,20 +72,24 @@ AzurePipelinesCredential::AzurePipelinesCredential(
             + "' needed by " + GetCredentialName() + ". This should be set by Azure Pipelines.");
   }
 
-  if (isTenantIdValid && !clientId.empty() && !serviceConnectionId.empty()
-      && !systemAccessToken.empty() && !m_oidcRequestUrl.empty())
+  if (TenantIdResolver::IsValidTenantId(tenantId) && !clientId.empty()
+      && !serviceConnectionId.empty() && !systemAccessToken.empty() && !m_oidcRequestUrl.empty())
   {
-    m_tokenCredentialImpl = std::make_unique<TokenCredentialImpl>(options);
-    m_requestBody
-        = std::string(
-              "grant_type=client_credentials"
-              "&client_assertion_type="
-              "urn%3Aietf%3Aparams%3Aoauth%3Aclient-assertion-type%3Ajwt-bearer" // cspell:disable-line
-              "&client_id=")
-        + Url::Encode(clientId);
+    ClientAssertionCredentialOptions clientAssertionCredentialOptions{};
+    // Get the options from the base class (including ClientOptions).
+    static_cast<Core::Credentials::TokenCredentialOptions&>(clientAssertionCredentialOptions)
+        = options;
+    clientAssertionCredentialOptions.AuthorityHost = options.AuthorityHost;
+    clientAssertionCredentialOptions.AdditionallyAllowedTenants
+        = options.AdditionallyAllowedTenants;
 
-    IdentityLog::Write(
-        IdentityLog::Level::Informational, GetCredentialName() + " was created successfully.");
+    std::function<std::string(Context const&)> callback
+        = [this](Context const& context) { return GetAssertion(context); };
+
+    // ClientAssertionCredential validates the tenant ID, client ID, and assertion callback and logs
+    // warning messages otherwise.
+    m_clientAssertionCredentialImpl = std::make_unique<_detail::ClientAssertionCredentialImpl>(
+        GetCredentialName(), tenantId, clientId, callback, clientAssertionCredentialOptions);
   }
   else
   {
@@ -140,6 +115,9 @@ Request AzurePipelinesCredential::CreateOidcRequestMessage() const
   request.SetHeader("content-type", "application/json");
   request.SetHeader("authorization", "Bearer " + m_systemAccessToken);
 
+  // Prevents the service from responding with a redirect HTTP status code (useful for automation).
+  request.SetHeader("X-TFS-FedAuthRedirect", "Suppress");
+
   return request;
 }
 
@@ -157,8 +135,21 @@ std::string AzurePipelinesCredential::GetOidcTokenResponse(
         + std::to_string(static_cast<std::underlying_type<decltype(statusCode)>::type>(statusCode))
         + " (" + response->GetReasonPhrase()
         + ") response from the OIDC endpoint. Check service connection ID and Pipeline "
-          "configuration.\n\n"
-        + responseBody;
+          "configuration";
+
+    auto responseHeaders = response->GetHeaders();
+    auto headerValue = responseHeaders.find("x-vss-e2eid");
+    if (headerValue != responseHeaders.end())
+    {
+      message += "\n" + headerValue->first + ":" + headerValue->second;
+    }
+    headerValue = responseHeaders.find("x-msedge-ref");
+    if (headerValue != responseHeaders.end())
+    {
+      message += "\n" + headerValue->first + ":" + headerValue->second;
+    }
+    message += "\n\n" + responseBody;
+
     IdentityLog::Write(IdentityLog::Level::Verbose, message);
 
     throw AuthenticationException(message);
@@ -194,7 +185,7 @@ AzurePipelinesCredential::~AzurePipelinesCredential() = default;
 std::string AzurePipelinesCredential::GetAssertion(Context const& context) const
 {
   Azure::Core::Http::Request oidcRequest = CreateOidcRequestMessage();
-  std::unique_ptr<RawResponse> response = m_httpPipeline.Send(oidcRequest, context);
+  std::unique_ptr<RawResponse> response = m_httpPipeline->Send(oidcRequest, context);
 
   if (!response)
   {
@@ -214,7 +205,7 @@ AccessToken AzurePipelinesCredential::GetToken(
     TokenRequestContext const& tokenRequestContext,
     Context const& context) const
 {
-  if (!m_tokenCredentialImpl)
+  if (!m_clientAssertionCredentialImpl)
   {
     auto const AuthUnavailable = GetCredentialName() + " authentication unavailable. ";
 
@@ -226,41 +217,6 @@ AccessToken AzurePipelinesCredential::GetToken(
         AuthUnavailable + "Azure Pipelines environment is not set up correctly.");
   }
 
-  auto const tenantId = TenantIdResolver::Resolve(
-      m_clientCredentialCore.GetTenantId(),
-      tokenRequestContext,
-      m_clientCredentialCore.GetAdditionallyAllowedTenants());
-
-  auto const scopesStr
-      = m_clientCredentialCore.GetScopesString(tenantId, tokenRequestContext.Scopes);
-
-  // TokenCache::GetToken() and m_tokenCredentialImpl->GetToken() can only use the lambda
-  // argument when they are being executed. They are not supposed to keep a reference to lambda
-  // argument to call it later. Therefore, any capture made here will outlive the possible time
-  // frame when the lambda might get called.
-  return m_tokenCache.GetToken(scopesStr, tenantId, tokenRequestContext.MinimumExpiration, [&]() {
-    return m_tokenCredentialImpl->GetToken(context, false, [&]() {
-      auto body = m_requestBody;
-      if (!scopesStr.empty())
-      {
-        body += "&scope=" + scopesStr;
-      }
-
-      // Get the request url before calling GetAssertion to validate the authority host scheme.
-      // This is to avoid making a request to the OIDC endpoint if the authority host scheme is
-      // invalid.
-      auto const requestUrl = m_clientCredentialCore.GetRequestUrl(tenantId);
-
-      const std::string assertion = GetAssertion(context);
-
-      body += "&client_assertion=" + Azure::Core::Url::Encode(assertion);
-
-      auto request
-          = std::make_unique<TokenCredentialImpl::TokenRequest>(HttpMethod::Post, requestUrl, body);
-
-      request->HttpRequest.SetHeader("Host", requestUrl.GetHost());
-
-      return request;
-    });
-  });
+  return m_clientAssertionCredentialImpl->GetToken(
+      GetCredentialName(), tokenRequestContext, context);
 }