+++ sdk/storage/azure-storage-common/inc/azure/storage/common/internal/xml_wrapper.hpp
@@ -4,6 +4,7 @@
 #pragma once
 
 #include <cstdint>
+#include <memory>
 #include <string>
 
 namespace Azure { namespace Storage { namespace _internal {
@@ -40,19 +41,15 @@ namespace Azure { namespace Storage { namespace _internal {
     explicit XmlReader(const char* data, size_t length);
     XmlReader(const XmlReader& other) = delete;
     XmlReader& operator=(const XmlReader& other) = delete;
-    XmlReader(XmlReader&& other) noexcept { *this = std::move(other); }
-    XmlReader& operator=(XmlReader&& other) noexcept
-    {
-      m_context = other.m_context;
-      other.m_context = nullptr;
-      return *this;
-    }
+    XmlReader(XmlReader&& other) noexcept;
+    XmlReader& operator=(XmlReader&& other) noexcept;
     ~XmlReader();
 
     XmlNode Read();
 
   private:
-    void* m_context = nullptr;
+    struct XmlReaderContext;
+    std::unique_ptr<XmlReaderContext> m_context;
   };
 
   class XmlWriter final {
@@ -60,13 +57,8 @@ namespace Azure { namespace Storage { namespace _internal {
     explicit XmlWriter();
     XmlWriter(const XmlWriter& other) = delete;
     XmlWriter& operator=(const XmlWriter& other) = delete;
-    XmlWriter(XmlWriter&& other) noexcept { *this = std::move(other); }
-    XmlWriter& operator=(XmlWriter&& other) noexcept
-    {
-      m_context = other.m_context;
-      other.m_context = nullptr;
-      return *this;
-    }
+    XmlWriter(XmlWriter&& other) noexcept;
+    XmlWriter& operator=(XmlWriter&& other) noexcept;
     ~XmlWriter();
 
     void Write(XmlNode node);
@@ -74,7 +66,8 @@ namespace Azure { namespace Storage { namespace _internal {
     std::string GetDocument();
 
   private:
-    void* m_context = nullptr;
+    struct XmlWriterContext;
+    std::unique_ptr<XmlWriterContext> m_context;
   };
 
 }}} // namespace Azure::Storage::_internal