+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -1,14 +1,26 @@
 # Release History
 
-## 12.13.0-beta.1 (Unreleased)
+## 12.13.0 (2024-09-17)
 
 ### Features Added
 
-### Breaking Changes
+- Features in `12.13.0-beta.1` are now generally available.
 
-### Bugs Fixed
+## 12.13.0-beta.1 (2024-08-07)
+
+### Features Added
+
+- Added ability to retrieve SAS string to sign for debugging purposes.
 
 ### Other Changes
+
+- [[#5765]](https://github.com/Azure/azure-sdk-for-cpp/pull/5765) Fix libcxx build. (A community contribution, courtesy of _[rschu1ze](https://github.com/rschu1ze)_)
+
+### Acknowledgments
+
+Thank you to our developer community members who helped to make Azure Storage Blobs better with their contributions to this release:
+
+- Robert Schulze _([GitHub](https://github.com/rschu1ze))_
 
 ## 12.12.0 (2024-07-16)